# perf: use reduce for chinese normalization